# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "在家办公小程序接口文档",
        "version": "1.0"
    },
    "paths": {
        "/api/activity/getlist": {
            "get": {
                "tags": [
                    "活动"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\ActivityController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/activity/getdetail": {
            "get": {
                "tags": [
                    "活动"
                ],
                "summary": "获取活动详情",
                "operationId": "App\\Http\\Controllers\\Api\\ActivityController::getDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "活动id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回详情"
                    }
                }
            }
        },
...
        "/api/articlecomment/getlist": {
            "get": {
                "tags": [
                    "文章管理"
                ],
                "summary": "获取文章评论列表",
                "operationId": "App\\Http\\Controllers\\Api\\ArticleCommentController::getList",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "article_id",
                        "in": "query",
                        "description": "文章id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "用户id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回所有已精选的评论及当前登录用户的所有评论，包含评论点赞数total"
                    }
                }
            }
        },
...
        "/api/article/getdetail": {
            "get": {
                "tags": [
                    "文章管理"
                ],
                "summary": "获取文章详情",
                "operationId": "App\\Http\\Controllers\\Api\\ArticleController::getDetail",
                "parameters": [
                    {
                        "name": "column_id",
                        "in": "query",
                        "description": "栏目id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回文章详情，包含评论总数、点赞总数、浏览总数"
                    }
                }
            }
        },
...
        "/api/column/getlist": {
            "get": {
                "tags": [
                    "栏目管理"
                ],
                "summary": "获取栏目",
                "operationId": "App\\Http\\Controllers\\Api\\ColumnController::getList",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "父id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "用户id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回所有栏目"
                    }
                }
            }
        },
...
        "/api/feedbacktype/getlist": {
            "get": {
                "tags": [
                    "个人中心"
                ],
                "summary": "获取意见反馈问题类型",
                "operationId": "App\\Http\\Controllers\\Api\\FeedbackTypeController::getList",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "返回所有"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/forumcomment/getlist": {
            "get": {
                "tags": [
                    "中心广场"
                ],
                "summary": "获取评论列表",
                "operationId": "App\\Http\\Controllers\\Api\\ForumCommentController::getList",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "forum_id",
                        "in": "query",
                        "description": "帖子id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "用户id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回所有已精选的评论及当前登录用户的所有评论，包含评论点赞数total"
                    }
                }
            }
        },
...
        "/api/forum/getdetail": {
            "get": {
                "tags": [
                    "中心广场"
                ],
                "summary": "获取详情",
                "operationId": "App\\Http\\Controllers\\Api\\ForumController::getDetail",
                "parameters": [
                    {
                        "name": "column_id",
                        "in": "query",
                        "description": "栏目id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回文章详情，包含评论总数、点赞总数、浏览总数"
                    }
                }
            }
        },
        "/api/forum/getlist": {
            "get": {
                "tags": [
                    "中心广场"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\ForumController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/news/getlist": {
            "get": {
                "tags": [
                    "新闻"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\NewsController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/news/getdetail": {
            "get": {
                "tags": [
                    "新闻"
                ],
                "summary": "获取详情",
                "operationId": "App\\Http\\Controllers\\Api\\NewsController::getDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回文章详情，包含评论总数"
                    }
                }
            }
        },
...
        "/api/question/getlist": {
            "get": {
                "tags": [
                    "问卷"
                ],
                "summary": "获取试题",
                "operationId": "App\\Http\\Controllers\\Api\\QuestionController::getList",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "返回问卷试题"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/quwenqiushi/getdetail": {
            "get": {
                "tags": [
                    "趣闻糗事"
                ],
                "summary": "获取详情",
                "operationId": "App\\Http\\Controllers\\Api\\QuwenQiushiController::getDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回文章详情，包含评论总数、点赞总数、浏览总数"
                    }
                }
            }
        },
        "/api/quwenqiushi/getlist": {
            "get": {
                "tags": [
                    "趣闻糗事"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\QuwenQiushiController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功,type=1表示趣闻，2表示糗事"
                    }
                }
            }
        },
...
        "/api/totormessage/getlist": {
            "get": {
                "tags": [
                    "导师"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\TutorMessageController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "分组id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/usermessage/getlist": {
            "get": {
                "tags": [
                    "用户管理"
                ],
                "summary": "获取分页列表",
                "operationId": "App\\Http\\Controllers\\Api\\UserMessageController::getList",
                "parameters": [
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "类型，1表示系统消息，2表示互动消息",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}